requirements.txt
appdirs==1.4.4
audioread==3.0.0
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
decorator==5.1.1
Flask==2.3.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
lazy_loader==0.2
librosa==0.10.0.post2
llvmlite==0.40.1rc1
MarkupSafe==2.1.3
msgpack==1.0.5
numba==0.57.0
numpy==1.24.3
packaging==23.1
pooch==1.6.0
pycparser==2.21
pydub==0.25.1
python-dotenv==1.0.0
requests==2.31.0
scipy==1.10.1
soundfile==0.12.1
soxr==0.3.5
threadpoolctl==3.1.0
typing_extensions==4.6.3
urllib3==2.0.3
Werkzeug==2.3.6
